+++ Contents/mods/Evolving Traits World/42.19/media/lua/shared/ETW_MarkDynamicTraits.lua
@@ -38,6 +38,7 @@ local function registerDynamicTraits()
 		ETW_Registry.traits.KNIFE_FIGHTER,
 		ETW_Registry.traits.LIGHTSTEP,
 		ETW_Registry.traits.LOW_PROFILE,
+		ETW_Registry.traits.NATURAL_EATER,
 		ETW_Registry.traits.OLYMPIAN,
 		ETW_Registry.traits.PAIN_TOLERANCE,
 		ETW_Registry.traits.PET_THERAPY,