+++ Contents/mods/Evolving Traits World/42.19/media/lua/shared/Types/ETW_ModDataType.lua
@@ -20,6 +20,7 @@
 --- @field EatingSpeedSystemCounter number
 --- @field HoarderCounter number
 --- @field OlympianCounter number
+--- @field NaturalEaterFoodsEaten integer
 --- @field HardyReserve number
 --- @field QuickRestLastEndurance number
 --- @field IdealWeightCounter number